Move player.schema.ts to the schemas folder
