+++ src/__tests__/player/data/player/playerBuilder.ts
@@ -1,5 +1,5 @@
 import { GameStatistics } from '../../../../player/gameStatistics.schema';
-import { Player } from '../../../../player/player.schema';
+import { Player } from '../../../../player/schemas/player.schema';
 import { ObjectId } from 'mongodb';
 
 export default class PlayerBuilder {
