+++ src/box/dto/createdBox.dto.ts
@@ -1,4 +1,4 @@
-import { Player } from '../../player/player.schema';
+import { Player } from '../../player/schemas/player.schema';
 import { Clan } from '../../clan/clan.schema';
 import { Chat } from '../../chat/chat.schema';
 import { Expose, Type } from 'class-transformer';
