+++ src/box/payloads/CreatedBox.ts
@@ -1,5 +1,5 @@
 import { Box } from '../schemas/box.schema';
-import { Player } from '../../player/player.schema';
+import { Player } from '../../player/schemas/player.schema';
 import { Clan } from '../../clan/clan.schema';
 import { Chat } from '../../chat/chat.schema';
