+++ src/player/decorator/validation/IsPlayerExists.decorator.ts
@@ -6,7 +6,7 @@ import {
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { Player } from '../../player.schema';
+import { Player } from '../../schemas/player.schema';
 import { registerValidationDecorator } from '../../../common/decorator/validation/registerValidationDecorator';
 import { isEntityExists } from '../../../common/decorator/validation/isEntityExists';
 import { ModelName } from '../../../common/enum/modelName.enum';
